+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.2.0] - 2025-01-03
+* ENH: Bulk Edit - Overwrite categories when edited, not appending new ones.
+* BUG: I18n loading fix.
+
 ## [1.1.0] - 2025-01-02
 * ENH: Abstract out max_marine_product_categories_enhancements_get_legacy_categories_ids.
 * ENH: New function max_marine_product_categories_enhancements_filter_legacy_categories_from_term_ids.